# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: ForceFlow API — Multi-Tenant Salesforce Automation Platform
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /chat:
    get:
      operationId: listThreads
      tags: [chat]
      summary: List chat threads for the current user
      responses:
        "200":
          description: Threads
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ChatThread"
    post:
      operationId: createThread
      tags: [chat]
      summary: Create a new chat thread
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChatThreadInput"
      responses:
        "201":
          description: Created thread
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ChatThread"
  /chat/{threadId}:
    get:
      operationId: getThread
      tags: [chat]
      summary: Get a specific thread with messages
      parameters:
        - name: threadId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Thread with messages
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ChatThreadWithMessages"
    delete:
      operationId: deleteThread
      tags: [chat]
      summary: Delete a chat thread
      parameters:
        - name: threadId
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Deleted
  /chat/{threadId}/messages:
    post:
      operationId: sendMessage
      tags: [chat]
      summary: Send a message in a thread (triggers AI response)
      parameters:
        - name: threadId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MessageInput"
      responses:
        "200":
          description: AI response (streaming)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ChatMessage"
  /chat/{threadId}/title:
    patch:
      operationId: updateThreadTitle
      tags: [chat]
      summary: Update thread title
      parameters:
        - name: threadId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ThreadTitleUpdate"
      responses:
        "200":
          description: Updated thread
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ChatThread"

  # Tool approvals
  /chat/{threadId}/approvals/{approvalId}:
    post:
      operationId: respondToApproval
      tags: [chat]
      summary: Approve or reject a tool-call approval request
      parameters:
        - name: threadId
          in: path
          required: true
          schema:
            type: string
        - name: approvalId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ApprovalResponseInput"
      responses:
        "200":
          description: Approval response recorded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ToolApproval"

  # ─── AGENT ANALYSES ─────────────────────────────────────────────────────────
  /analyses:
    get:
      operationId: listAnalyses
      tags: [analyses]
      summary: List agent analyses for the tenant
      parameters:
        - name: orgId
          in: query
          schema:
            type: string
        - name: status
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Analyses
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AgentAnalysis"
  /analyses/{analysisId}:
    get:
      operationId: getAnalysis
      tags: [analyses]
      summary: Get a specific analysis with full plan
      parameters:
        - name: analysisId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Analysis
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AgentAnalysis"
  /analyses/{analysisId}/approve:
    post:
      operationId: approveAnalysis
      tags: [analyses]
      summary: Approve an analysis plan for execution
      parameters:
        - name: analysisId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Approved analysis
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AgentAnalysis"

  # ─── PIPELINE RUNS ──────────────────────────────────────────────────────────
...
components:
  schemas:
...
    ChatThread:
      type: object
      required: [id, tenantId, userId, role, createdAt]
      properties:
        id:
          type: string
        tenantId:
          type: string
        userId:
          type: string
        title:
          type: ["string", "null"]
        role:
          type: string
          enum: [frontier, developer, sales, support, sme]
        orgId:
          type: ["string", "null"]
        modelId:
          type: ["string", "null"]
        messageCount:
          type: integer
        createdAt:
          type: string
          format: date-time
    ChatThreadInput:
      type: object
      required: [role]
      properties:
        role:
          type: string
          enum: [frontier, developer, sales, support, sme]
        orgId:
          type: string
        modelId:
          type: string
        title:
          type: string
    ThreadTitleUpdate:
      type: object
      required: [title]
      properties:
        title:
          type: string
    ChatThreadWithMessages:
      type: object
      required: [thread, messages]
      properties:
        thread:
          $ref: "#/components/schemas/ChatThread"
        messages:
          type: array
          items:
            $ref: "#/components/schemas/ChatMessage"
    ChatMessage:
      type: object
      required: [id, threadId, role, createdAt]
      properties:
        id:
          type: string
        threadId:
          type: string
        role:
          type: string
          enum: [user, assistant, tool]
        content:
          type: ["string", "null"]
        parts:
          type: ["object", "null"]
        toolCallId:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
    MessageInput:
      type: object
      required: [content]
      properties:
        content:
          type: string
        orgId:
          type: string
    ApprovalResponseInput:
      type: object
      required: [decision]
      properties:
        decision:
          type: string
          enum: [approved, rejected]
    ToolApproval:
      type: object
      required: [id, tenantId, threadId, toolName, status, createdAt]
      properties:
        id:
          type: string
        tenantId:
          type: string
        threadId:
          type: string
        messageId:
          type: ["string", "null"]
        toolName:
          type: string
        args:
          type: object
        status:
          type: string
          enum: [pending, approved, rejected]
        approvedBy:
          type: ["string", "null"]
        approvedAt:
          type: ["string", "null"]
          format: date-time
        createdAt:
          type: string
          format: date-time

    # ─── AGENT ANALYSES ────────────────────────────────────────────────────
    AgentAnalysis:
      type: object
      required: [id, tenantId, threadId, orgId, requirementSummary, status, createdAt]
      properties:
        id:
          type: string
        tenantId:
          type: string
        threadId:
          type: string
        orgId:
          type: string
        requirementSummary:
          type: string
        currentStateAnalysis:
          type: ["object", "null"]
        gapAnalysis:
          type: ["object", "null"]
        implementationPlan:
          type: ["object", "null"]
        status:
          type: string
          enum: [draft, approved, executing, done, failed]
        createdBy:
          type: ["string", "null"]
        approvedBy:
          type: ["string", "null"]
        approvedAt:
          type: ["string", "null"]
          format: date-time
        executionLog:
          type: ["object", "null"]
        createdAt:
          type: string
          format: date-time

    # ─── PIPELINES ─────────────────────────────────────────────────────────
...